ai-engine/requirements.txt
# ─────────────────────────────────────────
# WANDERIX — AI Engine Dependencies
# Python 3.11
# ─────────────────────────────────────────

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Anthropic Claude
anthropic==0.34.0

# Redis Cache
redis==5.0.8

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.8.2

# HTTP Client
httpx==0.27.2

# Async
anyio==4.4.0